==> Aspect Based Sentiment Analysis/Jupyter Notebooks/2. Get_Hotel_Reviews.ipynb <==
##########################################
# --- Importing the standard libraries ---
##########################################

import pandas as pd

################################
# --- Importing the Dataset ---
################################

data = pd.read_csv('../Prepared Data/Dataset.csv')
data.head()

---

################################################################
# --- Function to get reviews from city and docID of a hotel ---
################################################################

##-> Function Description
# Accepts a hotel doc_ID and city
# Returns a dataFrame with hotel reviews of that particular doc_ID and city


def getHotel_reviews(docID, city):
    reviews = pd.read_csv('../Raw Data/{0}/{1}'.format(city, docID), delimiter='\t', 
                          names=['Date', 'Review_Title', 'Review'], encoding='ISO-8859-14', 
                          usecols=range(3))
    return reviews

getHotel_reviews('china_beijing_holiday_inn_central_plaza', 'beijing')[100:105]

---

###############################################################
# --- Function to extract reviews and save it to a DataFrame ---
###############################################################

##-> Function Description
# It creates a list of combined reviews from each docID
# returns a DataFrame of Doc_ID and Review


def extract_reviews():
    
    review_list = []
    
    for city in list(data.City.unique()):
        for docID in list(data[data.City == city].Doc_ID):
            try:
                review = str(list(getHotel_reviews(docID, city)['Review']))
                review_list.append([docID, review])
            except:
                review_list.append([docID, 'No review'])
    
    df = pd.DataFrame(review_list, columns=['Doc_ID', 'Review'])

    return df


reviews_frame = extract_reviews() 

# len(reviews_frame)    # 2821 rows

---

###############################################################
#--- Merging the two data frames to prepare our final frame ---
###############################################################

ratings_frame = data[['Doc_ID', 'CLEANLINESS', 'ROOM', 'SERVICE', 'LOCATION', 'VALUE', 'OVERALL_RATING']]

combined_frame = pd.merge(reviews_frame, ratings_frame, on='Doc_ID', how='inner')

combined_frame.head()

---

####################################################################################
#--- Removing the Bad Records (where there is no review) and Saving the CSV file ---
####################################################################################


# len(combined_frame[combined_frame.Review == 'No review'])    # 1195 such rows to be dropped

combined_frame.drop(combined_frame[combined_frame.Review == 'No review'].index, inplace=True)

combined_frame.to_csv('../Prepared Data/Hotel_Reviews.csv', index=False)

# len(combined_frame)    1644 rows saved

---
