==> Aspect Based Sentiment Analysis/Jupyter Notebooks/3. Feature Engineering.ipynb <==
##########################################
# --- Importing the standard libraries ---
##########################################

import numpy as np
import pandas as pd
import re
import nltk
from textblob import TextBlob

################################
# --- Importing the Dataset ---
################################

data = pd.read_csv('../Prepared Data/Hotel_Reviews.csv')
data.head()

---

## Extracting the Features (Using Stopwords Removal and Lemmatization)

from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
from nltk.stem import WordNetLemmatizer

#########################################
# --- Function to Extract the Reviews ---
#########################################

##> Description
# Loops through all the reviews
# Apply's the following things:-
#        1. Removing all the symbols, dots, commas, etc. which are not required as features
#        2. Converting the alphabets to lower case
#        3. Splits the review into list of words
#        4. Removes the stopwords like pronouns, determiners, etc.
#        5. Appends all the clean reviews into a corpus list



def extract_features(data):
    corpus = []
    
    for i in range(0, len(data)):
        
        review = re.sub('[^a-zA-Z]', ' ', data['Review'][i])    
        review = review.lower()                                 
        review = review.split()                                 
        
        lmtzr = WordNetLemmatizer()
        review = [lmtzr.lemmatize(word) for word in review if not word in set(stopwords.words('english'))]
        
        review = ' '.join(review)
        
        corpus.append(review)
    return corpus


corpus = extract_features(data)

#########################
# --- Save our Corpus ---
#########################

with open('../Corpus/corpus.txt', 'a') as f:
    for line in corpus:
        f.write(line)
        f.write('\n')

---

## Extracting the Features (Using Parts of Speech (POS) Tagging)

#######################################
# --- Parts of Speech (POS) Tagging ---
#######################################
##> Extracting only Nouns and Adjectives which are the most describing words
# Appends a list of of tuples containing only 'NN' or 'JJ' into tag_list

def pos_extract(blob):
    tags_list = []
    for i in blob.tags:
        if (i[1] == 'NN') or (i[1] == 'JJ'):
            tags_list.append(i)
    return tags_list

#########################################
# --- Function to Extract the Reviews ---
#########################################

##> Description
# Loops through all the reviews
# Apply's the following things:-
#        1. Extract the POS tags for each review
#        2. Storing the nouns and adjectives into a review list
#        3. Appends all the extracted reviews into a corpus list



def extract_features_2(data):
    corpus = []
    
    for i in range(0, len(data)):
        
        review = []
        
        blob = TextBlob(data['Review'][i])
        
        NN_JJ_tags = pos_extract(blob)
        
        for word in NN_JJ_tags:
            review.append(word[0])
        
        review = ' '.join(review)
        
        corpus.append(review)
        
    return corpus


corpus = extract_features_2(data)

#########################
# --- Save our Corpus ---
#########################

with open('../Corpus/corpus2.txt', 'a') as f:
    for line in corpus:
        f.write(line)
        f.write('\n')

---

## Extracting the Features (Using Noun Phrases Extraction)

#################################
# --- Noun Phrases Extraction ---
#################################
##> Noun Phrase consists of (Det)(Adj)NN(PP)


def extract_features_3(data):
    corpus = []
    
    for i in range(0, len(data)):
        
        review = []
        
        blob = TextBlob(data['Review'][i])
        
        noun_phrase_list = list(blob.noun_phrases)
        
        for word in noun_phrase_list:
            review.append(word)
        
        review = ' '.join(review)
        
        corpus.append(review)
        
    return corpus


corpus = extract_features_3(data)

#########################
# --- Save our Corpus ---
#########################

with open('../Corpus/corpus3.txt', 'a') as f:
    for line in corpus:
        f.write(line)
        f.write('\n')

---

## Extracting the Features (Using Bi-Grams and N-Grams)

##############################
# --- Bi-Grams and N-Grams ---
##############################

# bi_grams = blob.ngrams(n=2)
# n_grams = blob.ngrams(n=3)

def extract_features_4(data, n):
    corpus = []
    
    for i in range(0, len(data)):
        
        review = []
        
        blob = TextBlob(data['Review'][i])
        
        bi_grams = blob.ngrams(n=n)
        
        for word in bi_grams:
            word = ' '.join(word)
            review.append(word)
        
        review = ' '.join(review)
        
        corpus.append(review)
        
    return corpus



corpus = extract_features_4(data, n=2)                # For Bi-Grams
# corpus = extract_features_4(data, n=3)                # For Tri-Grams

#########################
# --- Save our Corpus ---
#########################

with open('../Corpus/corpus4.txt', 'a') as f:
    for line in corpus:
        f.write(line)
        f.write('\n')

---
