==> Aspect Based Sentiment Analysis/Jupyter Notebooks/4. Modelling.ipynb <==
##########################################
# --- Importing the standard libraries ---
##########################################

import numpy as np
import pandas as pd

import warnings
warnings.filterwarnings('ignore')

################################
# --- Importing the Dataset ---
################################

data = pd.read_csv('../Prepared Data/Hotel_Reviews.csv')
data.head()

#############################################
#--- Loading the Corpus of Cleaned Reviews---
#############################################

corpus = []
with open('../Corpus/corpus.txt') as f:
    for line in f.readlines():
        corpus.append(line)

## Modelling

###############################################
# --- Chosing our Aspect for Classification ---
###############################################

# aspect = data.ROOM           # Accuracy Score:-> 60.331 %

# aspect = data.CLEANLINESS    # Accuracy Score:-> 58.197 %

# aspect = data.SERVICE        # Accuracy Score:-> 62.918 %

# aspect = data.LOCATION       # Accuracy Score:-> 61.169 %

# aspect = data.VALUE          # Accuracy Score:-> 59.866 %


# aspect = data.OVERALL_RATING        # Accuracy Scores:-> 70.213 % 

#########################################
# --- Creating the Bag of Words model ---
#########################################


from sklearn.feature_extraction.text import CountVectorizer
cv = CountVectorizer(max_features=5000)
X = cv.fit_transform(corpus).toarray()
y = list(map(int, aspect.round()))

##################################################################
# --- Splitting the dataset into the Training set and Test set ---
##################################################################

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, random_state=0)

########################################################################
# ---  Fitting Classifier to the Training set Predicting the results ---
########################################################################

from sklearn.naive_bayes import MultinomialNB

# Training our Classifier
classifier = MultinomialNB(alpha=0.00001)
classifier.fit(X_train, y_train)

# Predicting the Results
y_pred = classifier.predict(X_test)

# Computing the accuracy
from sklearn.metrics import accuracy_score
print("Accuracy Score:->", round(accuracy_score(y_test, y_pred) * 100,3), '%')


#---Accuracy Scores---

# MultinomialNB      # Accuracy Scores:-> 70.213 % , 65.653 % , 69.301 % , 71.125 %
# SVC                # Accuracy Scores:-> 65.35 %

##############################################################
# ---  Analyzing the Results using K-Fold Cross Validation ---
##############################################################

from sklearn.model_selection import cross_val_score
accuracies = cross_val_score(estimator = classifier, X = X_train, y = y_train, cv = 10)

mean_accuracy = accuracies.mean()
std_accuracy = accuracies.std()

print("Average Accuracy Score:->", round(mean_accuracy * 100,3), '%')
print("Standard Deviation:->", round(std_accuracy,3))

########################################################
# ---  Tuning the Hyper Parameters using Grid Search ---
########################################################

from sklearn.model_selection import GridSearchCV
# parameters = [{'alpha': [0.1,0.01,0.001]}]        # Best value: 0.001
# parameters = [{'alpha': [0.001,0.002,0.003]}]     # Best value: 0.001
# parameters = [{'alpha': [0.001,0.0001]}]          # Best value: 0.0001
# parameters = [{'alpha': [0.0001, 0.00001]}]       # Best value: 0.00001
parameters = [{'alpha': [0.000001, 0.00001]}]       # Best value: 0.00001 (Final Selected Value)

grid_search = GridSearchCV(estimator = classifier,
                           param_grid = parameters,
                           scoring = 'accuracy',
                           cv = 10,
                           n_jobs = -1)


grid_search = grid_search.fit(X_train, y_train)

best_accuracy = grid_search.best_score_
best_parameters = grid_search.best_params_

print("Best Accuracy Score:->", round(best_accuracy * 100,3), '%')
print("Best Parameters:->", best_parameters)

---
